# client/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "server | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5281/"
    }
  ],
  "paths": {
    "/api/auth/sign-up": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequestPOST"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequestPOST"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequestPOST"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestPOST"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestPOST"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestPOST"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationBaseResponsePOST"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mood": {
      "post": {
        "tags": [
          "Mood"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeMoodRequestPOST"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeMoodRequestPOST"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeMoodRequestPOST"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/user": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/update": {
      "patch": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "FullName": {
                    "type": "string"
                  },
                  "AvatarImage": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeMoodRequestPOST": {
        "required": [
          "feelings",
          "mood",
          "sleepHours",
          "moodDescription"
        ],
        "type": "object",
        "properties": {
          "feelings": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feeling"
            }
          },
          "mood": {
            "$ref": "#/components/schemas/MoodName"
          },
          "sleepHours": {
            "$ref": "#/components/schemas/SleepHours"
          },
          "moodDescription": {
            "type": "string"
          }
        }
      },
      "AuthenticationBaseResponsePOST": {
        "required": [
          "status",
          "jwt"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "jwt": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Feeling": {
        "enum": [
          "Overwhelmed",
          "Stressed",
          "Anxious",
          "Frustrated",
          "Disappointed",
          "Lonely",
          "Irritable",
          "Down",
          "Restless",
          "Tired",
          "Calm",
          "Content",
          "Peaceful",
          "Hopeful",
          "Motivated",
          "Grateful",
          "Optimistic",
          "Confident",
          "Excited",
          "Joyful"
        ]
      },
      "IdentityErrorCode": {
        "enum": [
          "PasswordRequiresNonAlphanumeric",
          "PasswordRequiresDigit",
          "PasswordRequiresLower",
          "PasswordRequiresUpper",
          "PasswordTooShort",
          "DuplicateUserName",
          "EmailIsInvalid",
          "UserNotFound"
        ]
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "LoginRequestPOST": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "MoodName": {
        "enum": [
          "VerySad",
          "Sad",
          "Neutral",
          "Happy",
          "VeryHappy"
        ]
      },
      "SignUpErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "$ref": "#/components/schemas/IdentityErrorCode"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SignUpRequestPOST": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SleepHours": {
        "enum": [
          "ZeroToTwo",
          "ThreeToFour",
          "FiveToSix",
          "SevenToEight",
          "NinePlus"
        ]
      },
      "UpdateUserResponse": {
        "required": [
          "status",
          "jwt"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "jwt": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "required": [
          "fullName",
          "email"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarURL": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UserStatus": {
        "enum": [
          "Inactive",
          "NoFullName",
          "Active"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Authentication"
    },
    {
      "name": "Mood"
    },
    {
      "name": "User"
    }
  ]
}